requirements.txt
numpy
django
djangorestframework
djangorestframework-simplejwt
groq
celery 
redis

